The Metabolite class

The Metabolite object is a representation of metabolomic data, metabolomic annotation, and sample annotation.

Value

A Metabolite class.

Slots

assayData

a data.frame or data.table of metabolite measurements (peak area data or normalized data, sample [row] * feature [column]).

featureData

a data.frame or data.table of metabolite annotation (chemical annotation)

sampleData

a data.frame or data.table of sample annotation (sample meta data).

featureID

a character of the metabolite ID column (in feature file and the column names of data), default: CHEM_ID (provided from Metabolon file).

sampleID

a character of the sample ID column (in sample and the first column of data), default: PARENT_SAMPLE_NAME (provided from Metabolon file).

logs

Log information of data analysis process.

miscData

Ancillary data.

See also