Create a Metabolite object

Create a Metabolite object from three input data sets: 1) metabolite measurements (eg. peak area data or normalized data), and 2) metabolite annotation (eg. chemical annotation) 3) sample annotation (eg. sample meta data)

Usage

create_Metabolite(
  assayData,
  featureData,
  sampleData,
  featureID,
  sampleID,
  logs
)

Arguments

assayData

a data.frame or data.table of metabolite measurements (peak area data or normalized data, sample [row] * feature [column]).

featureData

a data.frame or data.table of metabolite annotation (chemical annotation)

sampleData

a data.frame or data.table of sample annotation (sample meta data).

featureID

a character of the metabolite ID column (in feature file and the column names of data), default: CHEM_ID (provided from Metabolon file).

sampleID

a character of the sample ID column (in sample and the first column of data), default: PARENT_SAMPLE_NAME (provided from Metabolon file).

logs

Log information.

Value

A Metabolite object with slots: assayData, featureData, and sampleData. A Metabolite object.

See also

Examples

# df <- create_Metabolite(assayData = df_data, featureData = df_feature, sampleData =  df_sample)